fix: update API

+++ lib/screens/attendance/camera_screen.dart
@@ -1,7 +1,5 @@
 import 'dart:async';
-import 'dart:convert';
 import 'dart:io';
-import 'dart:typed_data';
 import 'package:faceshot_teacher/models/attendance.dart';
 import 'package:faceshot_teacher/models/teacher.dart';
 import 'package:faceshot_teacher/models/timetable.dart';
@@ -9,9 +7,11 @@ import 'package:faceshot_teacher/services/firebase_firestore_service.dart';
 import 'package:faceshot_teacher/services/network/attendance_predictor_api_client.dart.dart';
 import 'package:faceshot_teacher/utils/utility.dart';
 import 'package:firebase_storage/firebase_storage.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:camera/camera.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:image_picker/image_picker.dart';
 
 import 'attendance_screen.dart';
 
@@ -68,7 +68,11 @@ class _CameraScreenState extends State<CameraScreen> {
     return Scaffold(
       body: Stack(
         children: [
-          CameraPreview(controller!),
+          SizedBox(
+            height: MediaQuery.of(context).size.height,
+            width: MediaQuery.of(context).size.width,
+            child: CameraPreview(controller!),
+          ),
           Padding(
             padding: const EdgeInsets.only(bottom: 16),
             child: Align(
@@ -82,34 +86,76 @@ class _CameraScreenState extends State<CameraScreen> {
                         Text(currentProcess),
                       ],
                     )
-                  : IconButton(
-                      icon: const Icon(
-                        Icons.camera_rounded,
-                        size: 56,
-                      ),
-                      onPressed: () async {
-                        XFile? file = await takePicture();
-                        if (mounted) {
-                          if (file == null) {
-                            Utility.showSnackBar(
-                              context,
-                              'Please try again',
-                            );
-                          } else {
-                            //Get the picture
-                            XFile? imageFile = file;
+                  : Padding(
+                      padding: const EdgeInsets.only(bottom: 12),
+                      child: Row(
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: [
+                          //Camera
+                          IconButton(
+                            icon: const Icon(
+                              Icons.camera_rounded,
+                              size: 56,
+                            ),
+                            onPressed: () async {
+                              XFile? file = await takePicture();
+                              if (mounted) {
+                                if (file == null) {
+                                  Utility.showSnackBar(
+                                    context,
+                                    'Please try again',
+                                  );
+                                } else {
+                                  //Get the picture
+                                  XFile? imageFile = file;
 
-                            //Get the prediction and go to the next page
-                            setState(() {
-                              isLoading = true;
-                            });
-                            await _getAttendancePredication(imageFile);
-                            setState(() {
-                              isLoading = false;
-                            });
-                          }
-                        }
-                      },
+                                  //Get the prediction and go to the next page
+                                  setState(() {
+                                    isLoading = true;
+                                  });
+                                  await _getAttendancePredication(imageFile);
+                                  setState(() {
+                                    isLoading = false;
+                                  });
+                                }
+                              }
+                            },
+                          ),
+
+                          const SizedBox(width: 50),
+
+                          //Gallery
+                          IconButton(
+                            icon: const Icon(
+                              Icons.image_rounded,
+                              size: 56,
+                            ),
+                            onPressed: () async {
+                              //Get the picture
+                              final ImagePicker _picker = ImagePicker();
+                              XFile? imageFile = await _picker.pickImage(
+                                source: ImageSource.gallery,
+                              );
+
+                              //Get the prediction and go to the next page
+                              if (imageFile != null) {
+                                setState(() {
+                                  isLoading = true;
+                                });
+                                await _getAttendancePredication(imageFile);
+                                setState(() {
+                                  isLoading = false;
+                                });
+                              } else {
+                                Utility.showSnackBar(
+                                  context,
+                                  'No Image selected',
+                                );
+                              }
+                            },
+                          ),
+                        ],
+                      ),
                     ),
             ),
           )
@@ -144,22 +190,40 @@ class _CameraScreenState extends State<CameraScreen> {
     setState(() {
       currentProcess = 'Uploading picture to Cloud';
     });
-    Directory appDocDir = await getApplicationDocumentsDirectory();
+
+    late File fileAttendanceFinalImage;
+    late File? fileAttendanceCameraInitialImage;
     String attendanceUid =
         '${widget.selectedTimeSlot['from']}-${DateTime.now().month}-${DateTime.now().year}';
-    final fileAttendanceFinalImage =
-        File('${appDocDir.path}/$attendanceUid.jpg');
-    final fileAttendanceCameraInitialImage = File(classImage.path);
-    await fileAttendanceCameraInitialImage.copy(fileAttendanceFinalImage.path);
 
-    //Upload this image to Firebase Storage
-    TaskSnapshot snapshot = await FirebaseStorage.instance
-        .ref()
-        .child(
-            'Attendances/${widget.timetable.uid}/${fileAttendanceFinalImage.uri.pathSegments.last}')
-        .putFile(fileAttendanceFinalImage);
+    late TaskSnapshot snapshot;
+    if (kIsWeb) {
+      //path to save Storage
+      fileAttendanceFinalImage = File('temp/$attendanceUid.jpg');
+
+      //Upload this image to Firebase Storage
+      snapshot = await FirebaseStorage.instance
+          .refFromURL('urlFromStorage')
+          .child(
+              'Attendances/${widget.timetable.uid}/${fileAttendanceFinalImage.uri.pathSegments.last}')
+          .putFile(fileAttendanceFinalImage);
+    } else {
+      Directory appDocDir = await getApplicationDocumentsDirectory();
+      fileAttendanceFinalImage = File('${appDocDir.path}/$attendanceUid.jpg');
+      fileAttendanceCameraInitialImage = File(classImage.path);
+      await fileAttendanceCameraInitialImage
+          .copy(fileAttendanceFinalImage.path);
+
+      //Upload this image to Firebase Storage
+      snapshot = await FirebaseStorage.instance
+          .ref()
+          .child(
+              'Attendances/${widget.timetable.uid}/${fileAttendanceFinalImage.uri.pathSegments.last}')
+          .putFile(fileAttendanceFinalImage);
+    }
+
     if (snapshot.state == TaskState.success) {
-      // final String downloadUrl = await snapshot.ref.getDownloadURL();
+      final String downloadUrl = await snapshot.ref.getDownloadURL();
 
       //Get the Attendance from the ML model
       setState(() {
@@ -167,17 +231,17 @@ class _CameraScreenState extends State<CameraScreen> {
       });
 
       // Convert the image to base64
-      Uint8List uint8List = await classImage.readAsBytes();
-      String base64EncodedImage = base64Encode(uint8List);
-      List<Attendance> attendances =
-          await AttendancePredictorApiClient.getAttendancePrediction(
-        base64EncodedImage,
-      );
-
+      // Uint8List uint8List = await classImage.readAsBytes();
+      // String base64EncodedImage = base64Encode(uint8List);
       // List<Attendance> attendances =
       //     await AttendancePredictorApiClient.getAttendancePrediction(
-      //   downloadUrl,
+      //   base64EncodedImage,
       // );
+
+      List<Attendance> attendances =
+          await AttendancePredictorApiClient.getAttendancePrediction(
+        downloadUrl,
+      );
 
       //Update the Database
       setState(() {
@@ -190,7 +254,9 @@ class _CameraScreenState extends State<CameraScreen> {
       );
 
       //Delete the original file
-      fileAttendanceCameraInitialImage.delete();
+      if (fileAttendanceCameraInitialImage != null) {
+        fileAttendanceCameraInitialImage.delete();
+      }
 
       //Go to the next screen
       Navigator.push(
@@ -215,13 +281,5 @@ class _CameraScreenState extends State<CameraScreen> {
       });
       return;
     }
-
-    // Convert the image to base64
-    // Uint8List uint8List = await classImage.readAsBytes();
-    // String base64EncodedImage = base64Encode(uint8List);
-    // List<Attendance> attendances =
-    //     await AttendancePredictorApiClient.getAttendancePrediction(
-    //   base64EncodedImage,
-    // );
   }
 }
